+++ src/Exception/FailedToLoadFile.php
@@ -1,8 +1,0 @@
-<?php
-declare(strict_types = 1);
-
-namespace Innmind\Filesystem\Exception;
-
-final class FailedToLoadFile extends RuntimeException
-{
-}